main-application/requirements.txt
pyserial>=3.5
Pillow>=9.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
PyQt5>=5.15.0 